requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.4
PyJWT==2.8.0
requests==2.31.0